requirement.txt
google-adk
python-dotenv
mss
opencv-python
numpy